# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Облік Майна API",
    "version": "1.0"
  },
  "paths": {
    "/api/DELETE/main-asset/{id}": {
      "delete": {
        "summary": "Delete selected main asset",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "description": "Delete selected main asset",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the main asset to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Wrong query to database"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/DELETE/main-assets": {
      "delete": {
        "summary": "Delete selected main assets",
        "description": "Delete selected main assets",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids of the main assets to delete",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Wrong query to database"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/GET/document/{operation}/{id}/scans": {
      "get": {
        "summary": "Returns list of scan filenames",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "parameters": [
          {
            "name": "operation",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "string representation of the operation"
          },
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the document to get"
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "title": "ScanList",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/GET/document/{operation}/types": {
      "get": {
        "summary": "Returns list of types of documents",
        "description": "Returns list of document types by specific document operation",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON list of document types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schemas/api-get-document-{operation}-types.yaml#/DocumentType"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/GET/document/add-supply": {
      "get": {
        "summary": "Returns document AddSupply by type, number and date",
        "description": "Returns document AddSupply by type, number and date or null",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "allowedOperation id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "query",
            "description": "document number",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "document date",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON document AddSupply",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/api-get-document-addsupply.yaml#/DocumentAddSupply"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/GET/documents": {
      "get": {
        "summary": "Returns list of documents",
        "description": "Returns list of all documents to view",
        "security": [
          {
            "BearerAuth": [
              "bearer"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON list of all documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schemas/api-get-documents.yaml#/Document"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}